[pace] XML (split)
