+++ pace/exporter.py
@@ -1,93 +1,8 @@
 import xml.etree.ElementTree as ET
-from abc import abstractmethod, ABCMeta
 from xml.etree.ElementTree import Element, ElementTree
 
-from energyreport.classes.orientation import Orientation
 from info import DEBUG
-
-
-class PaceObject(metaclass=ABCMeta):
-    @property
-    @abstractmethod
-    def template_filename(self) -> str:
-        pass
-
-    @property
-    @abstractmethod
-    def path(self) -> str:
-        pass
-
-    @property
-    @abstractmethod
-    def replace_queries(self) -> dict:
-        pass
-
-
-class Roof(PaceObject):
-    template_filename = 'templates/roofplane.xml'
-    path = './building/skin/roofPlanes/INITIAL'
-
-    def __init__(self, name: str, orientation: Orientation, angle: float, width: float, height: float):
-        self.name = name
-        self.orientation = orientation
-        self.angle = angle
-        self.width = width
-        self.height = height
-
-    @property
-    def replace_queries(self):
-        return {
-            'shortDescription': {
-                'value': self.name
-            },
-            'orientation/INITIAL': {
-                'class': 'com.hemmis.mrw.pace.model.enums.Orientation',
-                'value': self.orientation.name
-            },
-            'slope/INITIAL': {
-                'class': 'java.math.BigDecimal',
-                'value': self.angle
-            },
-            'width/INITIAL': {
-                'class': 'java.math.BigDecimal',
-                'value': self.width
-            },
-            'height/INITIAL': {
-                'class': 'java.math.BigDecimal',
-                'value': self.height
-            }
-        }
-
-
-class Wall(PaceObject):
-    template_filename = 'templates/wallplane.xml'
-    path = './building/skin/wallPlanes/INITIAL'
-
-    def __init__(self, name: str, orientation: Orientation, width: float, height: float):
-        self.name = name
-        self.orientation = orientation
-        self.width = width
-        self.height = height
-
-    @property
-    def replace_queries(self) -> dict:
-        return {
-            'shortDescription': {
-                'value': self.name
-            },
-            'orientation/INITIAL': {
-                'class': 'com.hemmis.mrw.pace.model.enums.Orientation',
-                'value': self.orientation.name
-            },
-            'width/INITIAL': {
-                'class': 'java.math.BigDecimal',
-                'value': self.width
-            },
-            'height/INITIAL': {
-                'class': 'java.math.BigDecimal',
-                'value': self.height
-            }
-        }
+from .paceobjects.pace_object import PaceObject
 
 
 class PaceExporter:
